backend/pyproject.toml
[project]
name = "edulingo-backend"
version = "0.1.0"
description = "EduLingo AI v2 backend - FastAPI hexagonal"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "aioboto3>=12.3.0",
    "bcrypt>=4.2",
    "pyjwt[crypto]>=2.9",
    "python-multipart>=0.0.12",
    "mutagen>=1.47.0",
    "PyYAML>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "testcontainers[postgres,redis]>=4.8",
    "moto[s3]>=5.0.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: fast unit tests, no I/O",
    "integration: integration tests requiring services (postgres, redis, minio)",
    "e2e: end-to-end API tests via httpx AsyncClient",
]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
extend-per-file-ignores = { "alembic/versions/*.py" = ["I001"] }

[tool.ruff.lint.flake8-bugbear]
# FastAPI uses these as argument defaults by design.
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Path", "fastapi.Body", "fastapi.Header", "fastapi.Cookie", "fastapi.Form", "fastapi.File"]
